refactor(assistant): move DB template-cache logic into test helpers
Keep the test-only migration-template machinery out of the production
persistence module. `db-init.ts` now carries only a small seam: it reads a
`tryRestore()`/`save()` hook off `globalThis.vellumAssistant.dbTemplateCache`
and, when absent (always, in production), runs the real migrations unchanged.

The full implementation moves to `src/__tests__/db-template-helpers.ts` (path
hashing, VACUUM INTO snapshotting, WAL-safe restore, lazy-migration guard
resets). Because it is require()'d lazily on the first `initializeDb()` call —
at test-execution time, after the workspace override is set — it may import
`src/` like any regular test file; it is never on the preload's import chain.

A pure installer, `src/__tests__/db-template-cache.ts` (no static `src/`
imports), registers the hook via createRequire so the preload can load it
without pulling the drizzle/schema graph in at preload time. The test preload
calls `installDbTemplateCache()` alongside the other `install*()` mocks. This
mirrors the existing `dbSingletons` / `featureFlagCache` globalThis-slot
pattern, where the production owner and the test side reference the shared
namespace independently and neither imports the other.

Net: db-init.ts drops ~200 lines and no longer imports node:fs/crypto or the
test-only migration reset helpers. Behaviour is unchanged — full suite passes
except the same pre-existing environment-only failures; the cache still builds
and restores the four templates, and the un-awaited initializeDb() caller still
works because require() resolves synchronously within the sync prefix.
