+++ assistant/src/__tests__/test-preload.ts
@@ -22,6 +22,7 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterAll } from "bun:test";
 
+import { installDbTemplateCache } from "./db-template-cache.js";
 import { installGatewayIpcMock } from "./mock-gateway-ipc.js";
 
 // --- Phase 1: env override (zero source-module imports above this point) ---
@@ -52,6 +53,13 @@ delete process.env.CES_CREDENTIAL_URL;
 // retry loop in `initFeatureFlagOverrides()`. Tests that need specific IPC
 // responses use `mockGatewayIpc()` / `resetMockGatewayIpc()`.
 installGatewayIpcMock();
+
+// Install the DB migration-template cache so DB-touching test files restore a
+// pre-migrated template instead of re-running the full migration chain. The
+// real implementation is require()'d lazily on first use (see
+// `db-template-cache.ts`), so this call does not pull the persistence graph
+// into the preload's import chain.
+installDbTemplateCache();
 
 afterAll(() => {
   process.exitCode = 0;
